Allow players to drop their guns on death
Previously, we had to delete player guns on death because the endless
respawns ran the risk of performance issues or reaching max edict
limits.

With nt_drop plugin version 0.8.0+, we have the "sm_ntdrop_nodespawn"
cvar available for detecting whether weapons are despawned (and as such
can be safely dropped to the world by us).

This change allows for players to loot more guns and/or ammo from the
world, instead of eventually running out of bullets.